test: adjust test cases and code in general (wip)
